refactor playerblocks to autoextendclaimtask
